# docs/ApoxylTech_Phase1_OpenAPI.yaml
openapi: 3.1.0
info:
  title: ApoxylTech Innovations Hub — Phase 1 API
  version: "1.0.0"
  description: |
    Phase 1 MVP API for ApoxylTech Innovations Hub, covering four domains:
    Identity & Access, Client Management, Projects, and Website CMS.

    ## Scope
    ~25–30 endpoints, matching Section 6 of `ApoxylTech_Phase1_MVP_Spec.md`.
    Everything deferred there (invoicing, tickets, notifications, MFA, OAuth,
    Academy, Research, Innovation Lab, Analytics) has no endpoints here.

    ## Roles & Permissions
    Phase 1 has three roles — `admin`, `client`, `guest` — but authorization is
    enforced at the permission-code level (e.g. `project.read`, `project.update`,
    `blog.write`) so additional roles can be introduced later without changing
    this contract. Each operation below documents the permission(s) required.

    ## Security classification policy
    Every response schema in this spec follows the data dictionary's
    classification rules:
    - **Public / internal** fields may appear in responses.
    - **Sensitive / restricted** fields (password hashes, verification/reset
      token values, refresh token values after issuance, etc.) are never
      included in any response body, including admin-facing ones.
    Where a field exists on a table (per the ERD) but is excluded from every
    response schema for this reason, it is called out in a schema-level comment.
  contact:
    name: ApoxylTech Engineering
servers:
...
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user account
      security: []
      x-permission: none (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created; verification email sent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '409':
          description: Email already registered.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate with email and password
      security: []
      x-permission: none (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated. Returns access token in body and sets a secure, HttpOnly session/refresh cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          description: Invalid credentials, unverified email, or deactivated account.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Invalidate the current refresh token / session
      x-permission: authenticated (any role)
      responses:
        '204':
          description: Logged out.
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange a valid refresh token for a new access token
      security: []
      x-permission: valid refresh token (cookie)
      responses:
        '200':
          description: New access token issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          description: Refresh token missing, expired, or revoked.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/verify-email:
    post:
      tags: [Auth]
      summary: Confirm an email address using the token sent at registration
      security: []
      x-permission: none (public; possession of a valid, unexpired token acts as auth)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: >-
                    Single-use verification token. Note (security classification):
                    the `email_verifications.token` column is restricted and is
                    never returned by any GET/list endpoint — it only ever
                    reaches the user via the verification email itself.
      responses:
        '200':
          description: Email verified.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '400':
          description: Token invalid, expired, or already used.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password reset email
      security: []
      x-permission: none (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '202':
          description: >-
            Always returns 202 regardless of whether the email exists, to
            avoid leaking account existence.
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Set a new password using a reset token
      security: []
      x-permission: none (public; possession of a valid, unexpired token acts as auth)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token:
                  type: string
                  description: >-
                    Single-use reset token. Note: `password_resets.token` is
                    restricted; never returned by any response body.
                new_password:
                  type: string
                  format: password
                  minLength: 12
      responses:
        '200':
          description: Password updated. All existing refresh tokens for the user are revoked.
        '400':
          description: Token invalid, expired, or already used.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422':
          $ref: '#/components/responses/ValidationError'

  # ---------------------------------------------------------------- USERS
...
  /clients:
    get:
      tags: [Clients]
      summary: List clients
      x-permission: client.read (admin only in Phase 1)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Paginated list of clients.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /clients/{id}:
    get:
      tags: [Clients]
      summary: Get a single client
      x-permission: client.read (admin only in Phase 1)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Client detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------- PROJECTS
...
  /blog:
    get:
      tags: [Blog]
      summary: List published blog posts
      security: []
      x-permission: none (public) for published posts; blog.read for drafts, if requested by an authenticated admin
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Paginated list of blog posts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPostListResponse'
    post:
      tags: [Blog]
      summary: Create a blog post
      x-permission: blog.write (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlogPostCreateRequest'
      responses:
        '201':
          description: Blog post created (defaults to draft status).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPost'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
  /blog/{slug}:
    get:
      tags: [Blog]
      summary: Get a single published blog post by slug
      security: []
      x-permission: none (public) for published posts; blog.read to preview a draft
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Blog post detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPost'
        '404':
          $ref: '#/components/responses/NotFound'
  /blog/{id}:
    patch:
      tags: [Blog]
      summary: Update or publish a blog post
      x-permission: blog.write (admin only)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlogPostUpdateRequest'
      responses:
        '200':
          description: Updated blog post.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPost'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

  # ---------------------------------------------------------------- LEADS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token. Refresh token is delivered separately as a secure, HttpOnly, SameSite cookie — never in a JSON response body.
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    PageParam:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    PageSizeParam:
      name: page_size
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    ValidationError:
      description: Request failed schema validation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid access token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated, but lacking the required permission for this resource.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource does not exist or is outside the caller's scope.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Too many requests; rate limit applied to this endpoint.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: "validation_error" }
            message: { type: string }
            details:
              type: object
              additionalProperties: true
    PaginationMeta:
      type: object
      required: [page, page_size, total]
      properties:
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }

    # ------------------------------------------------------------ AUTH
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password:
          type: string
          format: password
          minLength: 12
          description: Hashed with Argon2 server-side; never stored or returned as plaintext.
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
    AuthTokenResponse:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token: { type: string, description: Short-lived JWT access token. }
        token_type: { type: string, enum: [bearer] }
        expires_in: { type: integer, description: Seconds until access token expiry. }
      description: >-
        Note (security classification): the refresh token itself is never
        present in this or any other JSON body — it is set only as a secure,
        HttpOnly cookie. `refresh_tokens.id`/expiry are internal bookkeeping,
        not exposed to clients.

    # ------------------------------------------------------------ USERS
    # users table columns: id, email, password_hash, status, created_at (+ audit fields)
    # password_hash is restricted and is excluded from every schema below.
    UserPublic:
      type: object
      description: Minimal representation returned right after registration, before the caller is authenticated.
      required: [id, email, status]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        status:
          type: string
          enum: [pending_verification, active, deactivated]
...
    Client:
      type: object
      required: [id, user_id, company_name]
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        company_name: { type: string }
        user:
          $ref: '#/components/schemas/UserPublic'
    ClientListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Client' }
        meta: { $ref: '#/components/schemas/PaginationMeta' }

    # ------------------------------------------------------------ PROJECTS
...
    BlogPostStatus:
      type: string
      enum: [draft, published]
    BlogPost:
      type: object
      required: [id, author_id, title, status]
      properties:
        id: { type: string, format: uuid }
        author_id: { type: string, format: uuid }
        title: { type: string }
        slug: { type: string }
        status: { $ref: '#/components/schemas/BlogPostStatus' }
        body: { type: string }
        published_at: { type: string, format: date-time, nullable: true }
    BlogPostCreateRequest:
      type: object
      required: [title, body]
      properties:
        title: { type: string }
        body: { type: string }
        status:
          $ref: '#/components/schemas/BlogPostStatus'
    BlogPostUpdateRequest:
      type: object
      properties:
        title: { type: string }
        body: { type: string }
        status: { $ref: '#/components/schemas/BlogPostStatus' }
    BlogPostListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/BlogPost' }
        meta: { $ref: '#/components/schemas/PaginationMeta' }

    # ------------------------------------------------------------ LEADS
    # leads table: contact-form PII, subject to NDPR retention policy flagged
    # in the MVP spec. Never joined into any other domain's response.
...